detect/http_user_agent: move tests into tests/
