+++ src/detect-engine-mpm.c
@@ -55,7 +55,6 @@
 #include "detect-engine-hmd.h"
 #include "detect-engine-hrhd.h"
 #include "detect-engine-hcd.h"
-#include "detect-engine-hua.h"
 #include "detect-engine-hhhd.h"
 #include "detect-engine-hrhhd.h"
 #include "detect-engine-hsmd.h"
