+++ src/runmode-unittests.c
@@ -45,7 +45,6 @@
 #include "detect-engine-hrud.h"
 #include "detect-engine-hsmd.h"
 #include "detect-engine-hscd.h"
-#include "detect-engine-hua.h"
 #include "detect-engine-hhhd.h"
 #include "detect-engine-hrhhd.h"
 #include "detect-engine-state.h"
@@ -209,7 +208,6 @@ static void RegisterUnittests(void)
     DetectEngineHttpRawUriRegisterTests();
     DetectEngineHttpStatMsgRegisterTests();
     DetectEngineHttpStatCodeRegisterTests();
-    DetectEngineHttpUARegisterTests();
     DetectEngineHttpHHRegisterTests();
     DetectEngineHttpHRHRegisterTests();
     DetectEngineInspectModbusRegisterTests();
